services/gps_verify/requirements.txt
postgres
pyzmq
